Create BFS.cpp
Breadth first search is one of the basic and essential searching algorithms on graphs.
As a result of how the algorithm works, the path found by breadth first search to any node is the shortest path to that node, i.e the path that contains the smallest number of edges in unweighted graphs. 
The algorithm works in O(N+M) time, where N is number of vertices and M is the number of edges.